+++ ext/polars/src/expr/rolling.rs
@@ -9,9 +9,10 @@ impl RbExpr {
         &self,
         window_size: usize,
         weights: Option<Vec<f64>>,
-        min_periods: usize,
+        min_periods: Option<usize>,
         center: bool,
     ) -> Self {
+        let min_periods = min_periods.unwrap_or(window_size);
         let options = RollingOptionsFixedWindow {
             window_size,
             weights,
@@ -45,9 +46,10 @@ impl RbExpr {
         &self,
         window_size: usize,
         weights: Option<Vec<f64>>,
-        min_periods: usize,
+        min_periods: Option<usize>,
         center: bool,
     ) -> Self {
+        let min_periods = min_periods.unwrap_or(window_size);
         let options = RollingOptionsFixedWindow {
             window_size,
             weights,
@@ -81,9 +83,10 @@ impl RbExpr {
         &self,
         window_size: usize,
         weights: Option<Vec<f64>>,
-        min_periods: usize,
+        min_periods: Option<usize>,
         center: bool,
     ) -> Self {
+        let min_periods = min_periods.unwrap_or(window_size);
         let options = RollingOptionsFixedWindow {
             window_size,
             weights,
@@ -117,9 +120,10 @@ impl RbExpr {
         &self,
         window_size: usize,
         weights: Option<Vec<f64>>,
-        min_periods: usize,
+        min_periods: Option<usize>,
         center: bool,
     ) -> Self {
+        let min_periods = min_periods.unwrap_or(window_size);
         let options = RollingOptionsFixedWindow {
             window_size,
             weights,
@@ -155,10 +159,11 @@ impl RbExpr {
         &self,
         window_size: usize,
         weights: Option<Vec<f64>>,
-        min_periods: usize,
+        min_periods: Option<usize>,
         center: bool,
         ddof: u8,
     ) -> Self {
+        let min_periods = min_periods.unwrap_or(window_size);
         let options = RollingOptionsFixedWindow {
             window_size,
             weights,
@@ -195,10 +200,11 @@ impl RbExpr {
         &self,
         window_size: usize,
         weights: Option<Vec<f64>>,
-        min_periods: usize,
+        min_periods: Option<usize>,
         center: bool,
         ddof: u8,
     ) -> Self {
+        let min_periods = min_periods.unwrap_or(window_size);
         let options = RollingOptionsFixedWindow {
             window_size,
             weights,
@@ -235,9 +241,10 @@ impl RbExpr {
         &self,
         window_size: usize,
         weights: Option<Vec<f64>>,
-        min_periods: usize,
+        min_periods: Option<usize>,
         center: bool,
     ) -> Self {
+        let min_periods = min_periods.unwrap_or(window_size);
         let options = RollingOptionsFixedWindow {
             window_size,
             min_periods,
@@ -273,9 +280,10 @@ impl RbExpr {
         interpolation: Wrap<QuantileInterpolOptions>,
         window_size: usize,
         weights: Option<Vec<f64>>,
-        min_periods: usize,
+        min_periods: Option<usize>,
         center: bool,
     ) -> Self {
+        let min_periods = min_periods.unwrap_or(window_size);
         let options = RollingOptionsFixedWindow {
             window_size,
             weights,