+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Aliased `melt` to `unpivot` for `DataFrame` and `LazyFrame`
 - Changed signature of `pivot` and `melt` methods
 - Changed signature of `date_range` and `date_ranges` methods
+- Removed `use_earliest` option from `replace_time_zone` and `to_datetime` methods (use `ambiguous` instead)
 - Removed `explode` method from `StringExpr` (use `split("").explode` instead)
 - Removed `set_ordering` method from `CatExpr`